requirements.txt
aiogram==3.21.0
aiohttp==3.11.10
python-dotenv==1.0.1
openai==1.58.1